Move duplicate-check code to subfolder
