+++ frontend/viewer/src/lib/entry-editor/duplicate-check/DuplicateCheck.svelte
