+++ frontend/viewer/src/lib/entry-editor/duplicate-check/DuplicateCheckSection.svelte
