+++ frontend/viewer/src/lib/entry-editor/duplicate-check/DuplicateSummaryPill.svelte
