+++ frontend/viewer/src/lib/entry-editor/duplicate-check/duplicate-check.test.ts
